Calendar: fix Occurred "com.android.calendar" force close.
Like commit() but allows the commit to be executed after an activity's
state is saved. This is dangerous because the commit can be lost if
the activity needs to later be restored from its state, so this should
only be used for cases where it is okay for the UI state to change
unexpectedly on the user.

Use DialogFragment's method showAllowingStateLoss instead of method show.
The method showAllowingStateLoss use commitAllowingStateLoss to add a
fragment.

Change-Id: I88d7b401484d31de1ca92ec13af19fcd470553a7
CRs-Fixed: 1080088